Python/DP/Creational/AbstractFactoryMethod/requirements.txt
PyYAML==6.0
mysql-connector-python==8.0.32